Separate media tests to another folder
